handle mobs leaving better
Sometimes your action (hide; backstab %1) takes a tick before it occurs.
This can cause a failure of the bot when a mob leaves during that tick.
Also, there are times when the bot arrives and the mob that just
triggered a fight is leaving.  Sitting inbattle back to 0 and sitting
wasinbattle to 1, keeps the bot from moving forward when it otherwise
wouldn't.